IMDP storage changes.
The IMDP interface no longer extends MDP<Interval>. This created
to many problems due to it extending both Model<Interval<Value>>
and Model<Value>, and it did not support many MDP methods anyway.
Instead, it extends a (new) UMDP interface, to which most old
IMDP methods have been moved.

There is a new UMDP model type, currently only only used in the
UMDP interface (and overwritten by IMDP). This will later be used
for other implementations of uncertain MDPs.

The IMDPSimple implementation of IMDP also changes. Rather than
extend an MDPSimple<Interval>, it uses one for internal storage.
In a related change, IMDPSimple<Value>'s evaluator is now also
of type <Value> and there is a separate setIntervalEvaluator()
method to provide an evaluator for intervals. As before, for the
usual case of doubles, everything just works.